feat: add VVC NAL unit support to mp4ff-nallister
  - Add VVC codec option (-c vvc) to command line interface
  - Change return type of Sample.PresentationTime() to int64
  - Implement VVC AUD detection for Annex B stream processing

+++ examples/multitrack/main.go
@@ -119,7 +119,7 @@ func writeScenaristFile(w io.Writer, clcpTrack *Track) error {
 		return err
 	}
 	for _, sample := range clcpTrack.samples {
-		tMs := sample.PresentationTime() * 1000 / clcpTrack.timeScale
+		tMs := uint64(sample.PresentationTime()) * 1000 / clcpTrack.timeScale
 		msg := timeFromMs(tMs)
 		buf := bytes.NewBuffer(sample.Data)
 		box, err := mp4.DecodeBox(0, buf)
